Use dry-core extensions lib for use specific colorize options in logger
