+++ lib/dry/monitor/sql/colorizers/default.rb
@@ -1,0 +1,16 @@
+module Dry
+  module Monitor
+    module SQL
+      module Colorizers
+        class Default
+          def initialize(_theme)
+          end
+
+          def call(string)
+            string
+          end
+        end
+      end
+    end
+  end
+end
