+++ lib/dry/monitor/sql/logger.rb
@@ -1,6 +1,6 @@
 require 'dry-configurable'
+require 'dry/core/extensions'
 require 'dry/monitor/notifications'
-require 'dry/monitor/sql/colorize'
 
 module Dry
   module Monitor
@@ -8,21 +8,38 @@ module Dry
 
     module SQL
       class Logger
+        extend Dry::Core::Extensions
         extend Dry::Configurable
 
+        register_extension(:default_colorizer) do
+          require_relative './colorizers/default'
+
+          def colorizer
+            @colorizer ||= Colorizers::Default.new(config.theme)
+          end
+        end
+
+        register_extension(:rouge_colorizer) do
+          require_relative './colorizers/rouge'
+
+          def colorizer
+            @colorizer ||= Colorizers::Rouge.new(config.theme)
+          end
+        end
+
         setting :theme, nil
-        setting :colorize, true
         setting :message_template, %(  Loaded %s in %sms %s).freeze
 
         attr_reader :config
         attr_reader :logger
         attr_reader :template
 
+        load_extensions(:default_colorizer)
+
         def initialize(logger, config = self.class.config)
           @logger = logger
           @config = config
           @template = config.message_template
-          @colorize = Colorize.new(config.theme)
         end
 
         def subscribe(notifications)
@@ -32,13 +49,7 @@ module Dry
         end
 
         def log_query(time, name, query)
-          logger.info template % [name.inspect, time, colorize(query)]
-        end
-
-        private
-
-        def colorize(string)
-          config.colorize ? @colorize.call(string) : string
+          logger.info template % [name.inspect, time, colorizer.call(query)]
         end
       end
     end
